requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Base de datos
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0         # driver PostgreSQL async
psycopg2-binary==2.9.9  # driver PostgreSQL sync (para Alembic)

# Supabase
supabase==2.4.2

# Autenticación
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Validación y settings
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1

# HTTP client
httpx==0.27.0

# Tareas async
celery==5.4.0
redis==5.0.4

# Storage / imágenes
pillow==10.3.0

# QR
qrcode[pil]==7.4.2

# Reportes
openpyxl==3.1.2
reportlab==4.1.0

# Utils
python-dotenv==1.0.1
python-slugify==8.0.4

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
